user privacy managing and dealing with AnonymousPatron new syspref
Add a new option in patron table that let the user decide how to deal with his reading history.

3 options are available :
* never remove my reading list (keep it forever)
* let the library decide (legally keep my reading list, the default value)
* immediatly remove my reading history when I return a book (don't keep any reading history at all)

the OpacPrivacy syspref let the library decide if this option is active or not.

This patch also creates a new syspref, AnonymousPatron, that contains the borrowernumber of the Patron to attach anonymised issues.
The existing AnonSuggestion is modified to become a YesNo.