[HEAD] (bug #3323) allow rules for reserves in issuing rules.
This patch move reserves to issuing rules, and check reserves from branchcode/categorycode/itemtype when it's possible.
In some case we cannot know the branchcode or the itemtype of the document needed to be reserved, so default values(in issuing rules) are used.
It introduce a new syspref : ReservesControlBranch that define which branch should be used to control the user reserves rights and a new column in issuingrules table to store reserves rules.