Ergonomy improvement in smart rule management
Added a jQuery table filter, usefull when you have a lot of rows. This filter works on the entire row and allows you for exemple to type "Adul Perio 5".
Indentation cleaning.
Added tabs to filter colums on 4 axes: All, Issues, Fines and Reserves.
Moved the table filter to the right of the tabs.