[MT2355] Smart rules rewritte
Created an issuing rule API, and removed the SQL queries in smart-rules.
Default rules are now read from the issuingrule table, this makes the default_* tables deprecated.
So I also replaced the subs that where using these tables. This patch needs a lot of tests.
Migrating old rules to issuingrule will be done in updatedatabase in a later patch.
We need to move exsisting rule related subs to our new API.