(MT #3006) rewrite of reserve rights management
The way to check if a user have right was to complicated and buggous, this patch made it as a different way.
Firstly we check a user have the right rights or not to hold, and then we check how much the category code can hold.

This patch fix some warnings too in request.pl