[MT2783] removing AutoCommit=0, the last item is not commited
For an unknown reason, the last item batch updated is NOT updated.
After investigating a LOT (more than 3 hours), it appears the problem comes from the AutoCommit=0
I tried to solve it, but failed. So I remove the AutoCommit=0 option.

this result in a loss of perfs, but it's not a big problem I think. OTH, now everything is updated !