MT5248 link biblio => serials subscription is wrong
The GetSubscription function can be called with 3 parameters : title, issn, biblionumber
In biblio view (both opac & staff interface), the link to jump from the biblio to the subscription(s) was done using title

The result was that for a title like "Science", all subscriptions containing "science" in the title were linked.

That's too much and that's false.

This commit fixes this problem by calling GetSubscription with biblionumber only where applicable