MT7654 Merge authorities when repeated (dedup script)
Squashed commits
----------------

AuthoritiesMarc merge fix. + reindenting.

If $9 was at the end of the biblio record,
the authority information were not deleted
Now it is the case

Debug merge in C4/Authorities.pm for solr

Adding script dedup_authorities.pl

dedup authorities for solr

Merge authorities : widening the synched fields on biblios

search was restricted to fields related to the authority code remote.
It is now looking into all the tags linked to any authoriy rather than looking only
to the tag linked to this authtypecode.

Adding support for multiple composed authorities

With field 606 in UNIMARC when composed authorities, you need to change
the subfield from the authority to insert them into subfields of the
biblio.
This process adds that function.
It changes the subfieldcode from $a to $y if biblio tag is 606 and authtypecode is linked to 215.
It changes the subfieldcode from $a to $z if biblio tag is 606 and authtority contains "chronologique".

Using Storable to store and retrieve the authority

Stores the old authority in the file generated when updating an authority.
This file will be read and used in merge_authority.pl

MT7337 followup : adding auth_tag_to_report

AuthoritiesMarc. merge :$3 were munged.

$3 when merging authorities into biblio records were munged.
This was therefore introducing a problem with vendangeur

MT7654 : merge fix

Problem on the squashing
Introduced a bogus merge.
Fixing the merge
fixing compile error