MT5671: fixed issue with 'Renew or return' button
In circ/circulation.pl, after exporting some notices, checking 'Renew'
or 'Check in' checkboxes and then clicking on 'Renew or Return' button,
the export script is run instead of the renew script.
This was due to a modification of the form action, which was not been
reset to its initial value.

Squashed commits:
MT5671: Enable sorting on 3 columns more
MT5671: Sorting problem in issues table
MT5671: Move a 'form' closing tag at the right place