MT6533: New page invoice.pl (Invoice details)
invoice.pl shows the list of orders that are related to an invoice.
On invoices.pl the user can close or reopen an invoice.
Closing an invoice cause to prevent user for receiving an order within
this invoice.
User can also close or open an invoice in parcel.pl.
When creating a new invoice, an error message is shown if the
invoicenumber already exists.
Added calendars for invoices search.
Renamed 'Vendor price' to 'Retail price (taxes included)' in
neworderempty.pl