MT6063 : addbiblio.pl : authorised values with 0
authorised value count 0 as a value would not be correctly processed.
Indeed, the CGIScrolling list would not be passed the correct value
because the test was not done on defined but on the value itself.

This adds a test on defined value in order to fix that.

Way to reproduce :
    - Add an authorised value category with authorised value 0 "mylib" as description and some others
    - Add a link to your framework to a subfield and change authorisedvalue in the more constraints on the category you edited
    - take a biblio, edit it.
    - Store the value 0
    - go to MARCdetail.pl you see it
    - click on Edit Record
    - one expects to see "mylib" displayed in the subfield linked to that category.
    But shows nothing and if saved as such, forgets the value
(cherry picked from commit 8791a45ea8076f8b32d49c106beea0449ff976ae)