MT6523: Display a unique item block at once
On orderreceive.pl, when AcqCreateItem is 'receiving', it displays an
item block with item infos to fill, and a '+' button.
When user clicks on '+', the block is hidden and a list shows up with
the items that will be received. User can then edit or delete items in
the list and click 'Save' to receive items.