Fixing PrepareItemRecordDisplay : defaultvalue management
The default values management in PrepareItemRecordDisplay had not been
updated with all the facilities added in acqui/addbiblio.pl
This patch fixes that.
I prepared a branch Item_Input_Refactoring
This branch is quite old... And therefore would need some overload to
get it back into the trunck.
And the same work should be done for Biblio.
I think this should be done on the community version rather than in our
master.