MT7202: Fix GetNextDate when periodicity is by week
When previous date was in the last week of a year, but in the beginning
of the next year (e.g. Sunday 01/01/2012) and periodicity is by week
(e.g. 1 issue per week), GetNextDate returned the date of monday of the
next week, but with the wrong year (e.g. 07/01/2013).
This patch fix that.