MT5508 Fix some bugs that broke display of z3950 search results
z3950 search results are displayed in tabs (one tab for each server,
and tabs are built in the same order than results for a server are
retrieved. The problem was search is made asynchronously and results
never comes in the same order.
This patch assign an order (based on 'rank' field) to servers in order
to always have the same display and avoid errors in passing parameters
in url (especially the 'tab' parameter).