Reserves with Global rules
This allows reserves rules as the circulation rules
It tests defined branch then global branch
and defined itemtype and global itemtype
and if the reserves count is more than reservesallowed
then it Blocks.

It should be noted that GetIssuingRule returns No rules i.e. no
rights.
The Issuing rules are defined as follow
      [ $categorycode,       $itemtype,       $branchcode ],
      [ $categorycode,       "Default",       $branchcode ],
      [ "Default",           $itemtype,       $branchcode ],
      [ "Default",           "Default",       $branchcode ],
      [ $categorycode,       $itemtype,       "Default" ],
      [ $categorycode,       "Default",       "Default" ],
      [ "Default",           $itemtype,       "Default" ],
      [ "Default",           "Default",       "Default" ],
And values are inherited from bottom to top if no other is defined
and takes precedence.

Reserves rules are tested as follow:
    branchcode,categorycode,itype
    branchcode,categorycode,*
    *,categorycode,itype
    *,categorycode,*
Where * is the Global rule
If any of those rule fails then it cancel the reserve