Some Simple Load Testing pages
those pages may help you profiling the server.
How so ?
Edit the files so as to change kohauser and password to what they should be in your context.
(That was meant in order to be out of any consideration of ... say C4::Context)
Then link the directory inside koha-tmpl

On a remote computer,
launch
for p (memcached.pl memoize_memcached.pl simplepage.pl sysprefs_fields.pl sysprefs_limited.pl tablesysprefs.pl)
do
    ab -n 500 -c 50 http://opac/cgi-bin/koha/load_testing/$p &>logs_$p
done

And you're done.
Be aware though that those are NOT true Koha pages, and they are much lighter.
But this is a nice start to test