MT6615: Fix authority summary building
Before this fix, we were looping through all fields and subfields to
find a match in the summary. Moreover, the field values were not placed
in the order the librarian wrote.
Now, we parse the summary from beginning to end, replacing tags with
the corresponding value, in one pass.
It is cleaner and much more efficient.
(cherry picked from commit aaa7087f3d1ada34e208b9c4bc3d92a081447142)

Squashed commit:

MT6615: Prevent BuildSummary to go into an infinite loop

When an authority contains a field or a subfield which contains a string like
[sometextXXXYsometext] where XXX is a three digit number and Y is a character
in [\*a-z0-9], BuildSummary tries to replace it with the right field value.
But if it don't succeed, it continues matching this pattern and always find
the same string, so it loops indefinitely.
This patch fixes that by saving first all the fields to replace, and then
replace them (only the first match), beginning from the last field in the
string.
(cherry picked from commit 4853ac715919518826c19dd906a1e6aaeca53e25)