MT6615: Fix authority summary building
Before this fix, we were looping through all fields and subfields to
find a match in the summary. Moreover, the field values were not placed
in the order the librarian wrote.
Now, we parse the summary from beginning to end, replacing tags with
the corresponding value, in one pass.
It is cleaner and much more efficient.