MT6879: Replace SQL search by SolR search
Search is done using indexes is-serial and srt_*_title, so to have this
working, you have to ensure that is-serial is a valid index for SolR and
is mapped to 099$s (which have to be mapped with biblio.serial), and
*_title index must be sortable.
If you want stopwords to not be considered in search, check the
stopwords table.
And, finally, be sure that biblio records are well indexed.

This patch also add pagination to opac-serials.pl

Note: I commented out a few lines in opac-serials.pl which serve to
disable links on letters which have no results. Doing this can be quite
slow, because it send a request for each letter (26) to check for
the number of results. That's why I commented it out.