MT6615: Prevent BuildSummary to go into an infinite loop
When an authority contains a field or a subfield which contains a string like
[sometextXXXYsometext] where XXX is a three digit number and Y is a character
in [\*a-z0-9], BuildSummary tries to replace it with the right field value.
But if it don't succeed, it continues matching this pattern and always find
the same string, so it loops indefinitely.
This patch fixes that by saving first all the fields to replace, and then
replace them (only the first match), beginning from the last field in the
string.