MT7562: Clean GetNextDate and fix irregularities
Subscriptions irregularities are saved in the form X;X;X;... where X is
the number of the issue which will not be received by the library.
GetNextDate is now fully functionnal with these irregularities

This patch includes some minor fixes in irregularities management
especially in showpredictionpattern.pl which use the new sub
GetFictiveIssueNumber to get the right issue number.
Subscription irregularities can now be modified, even after receiving
some issues, it should work...

When periodicity is N by week (or month or year) and N is greater than
1, all issues in the same week (or month or year) are given the same
date than the previous issue or the first day of the period if the is no
previous issue. The librarian has to modify it by himself, because we
can't guess what is the right date.