MT7562: Allow to skip issuenumber when an irregularity is found
New field subscription.skip_serialseq, modifiable in
subscription-add.pl. If on (=1), this causes GetNextSeq to skip as many
serialseq as irregularities found.
E.g. you have a serial issue with number 'Issue 4' published on
2011-08-26. The next issue is irregular and its planned date is
2011-08-27.
If skip_serialseq is off, you will have for the next issue:
 - serialseq : 'Issue 5'
 - publisheddate : '2011-08-28'
If skip_serialseq is on:
 - serialseq : 'Issue 6'
 - publisheddate : '2011-08-28'