Adding support for multiple composed authorities
With field 606 in UNIMARC when composed authorities, you need to change
the subfield from the authority to insert them into subfields of the
biblio.
This process adds that function.
It changes the subfieldcode from $a to $y if biblio tag is 606 and authtypecode is linked to 215.
It changes the subfieldcode from $a to $z if biblio tag is 606 and authtority contains "chronologique".