[SOLR] Fix deletion of RecordIndex
With this code (in DeleteRecordIndex):

    my $sc = GetSolrConnection;
    $sc->remove("id:${recordtype}_${id}");

WebService::Solr::delete fail if version of WebService::Solr is >= 0.14
causing this software error

    invalid XML data: character data isn't a string

The cause of this is that Data::SearchEngine::remove takes two
parameters and calls directly WebService::Solr::delete like this:

    $self->_solr->delete({ query => $param1, id => $param2 });

If $param2 is undefined, WebService::Solr tries to create an XML element
with this undefined value, causing XML::Easy::Element->new to fail.

With version 0.11 of WebService::Solr, it does not fail because it uses
another XML library (XML::Generator).

This patch simply provides an empty table ref to $sc->remove as 2nd
parameter, so it does not change the current behaviour and is
compatible with both versions of WebService::Solr