FIX: do not try to display inexistant biblio in catalogue/search.pl
C4::Search::getItemsInfos calls GetBiblioData with a biblionumber. If
this biblionumber no more exists in database, it returns 'undef', and
HTML::Template::Pro will fail to display the results table if there is
an undef value in the loop.
It causes some inconsistencies (display less records than the pagination
parameter) but they can be fixed by reindexing all biblios