Deployment for the Search Engine queue.
This adds a new line to koha-conf.xml
install_dir est à rajouter pour pointer vers le dossier d'install
It fetches the daemon configuration from a configuration file (a YAML file
in config/daemon/indexrecord/config.yaml)
This should be quite safe. But you NEED to add a line to the koha-conf.xml file
Be also aware that you would need to add
config/daemon/indexrecord/config.yaml to install_dir (usually /home/koha)
in order to be able to embed the configuration into the daemon.

This C4::Config::.. Is the first commit for a complete rewrite of C4::Context Using YAML and hierarchy