MT6863: Adds modifications in add patrons
Add 2 fields in borrowers table : endguaranteedate and sudoc_login

Add syspref BorrowerUnwantedField (based on a community dev BZ6190):
It's contains a string with a fields list. theses fields no appears in
add member form.

format:
   category1:field1|field2;field3;category2:field4

In this example, field1 and field2 no appear if patron is a category1.
field3 never appears.
field4 no appears if patron is a category2