MT8059 : Error 500 on member search
When searching on a base where there is no accountlines for users, causes an error 500

Can't use an undefined value as an ARRAY reference at /home/koha/src/C4/Members.pm
line 731.
This patch changes the way the DBI is invoked so that no error would be sent when no data.