MT6861 followup : Sideeffect on members list.
The previous development  introduced a side effect when no branch is
assigned to a librarian or when display of a list of members of a branch
different from the librarian's.
It resulted in an ERROR 500.
This patch is solving the problem by not getting the category display
names when the member branch is not the same as the librarian's.